requirements.txt
google-generativeai
python-dotenv
duckduckgo_search
flask
pytest